refactor(store): move store modules to a separate library
moved store modules to a separate library

BREAKING CHANGE: reinstall project dependencies
