+++ apps/nx-ng-starter/src/app/components/app/app.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, Inject, ViewChild } from '@angular/
 import { MatSidenav } from '@angular/material/sidenav';
 import { Title } from '@angular/platform-browser';
 import { APP_ENV, WebEnvironment } from '@nx-ng-starter/shared-core/data-access';
-import { SidebarUiService } from '@nx-ng-starter/shared-core/ui';
+import { SidebarUiService } from '@nx-ng-starter/shared-store/state';
 
 /**
  * Application root component.
