+++ functions/package.json
@@ -20,17 +20,17 @@
     "@nestjs/graphql": "7.4.6",
     "@nestjs/jwt": "7.0.0",
     "@nestjs/microservices": "7.2.0",
-    "@nestjs/passport": "7.0.0",
+    "@nestjs/passport": "7.1.0",
     "@nestjs/platform-express": "7.2.0",
-    "apollo-server-express": "2.15.0",
+    "apollo-server-express": "2.15.1",
     "class-transformer": "0.2.3",
     "class-validator": "0.12.2",
     "dotenv": "8.2.0",
     "express": "4.17.1",
-    "firebase-admin": "8.12.1",
+    "firebase-admin": "8.13.0",
     "firebase-functions": "3.7.0",
     "google-protobuf": "3.12.2",
-    "graphql": "15.1.0",
+    "graphql": "15.2.0",
     "graphql-subscriptions": "1.1.0",
     "graphql-tag": "2.10.3",
     "graphql-tools": "6.0.11",
