+++ libs/mocks-core/src/index.ts
@@ -1,6 +1,3 @@
-/**
- * Mocks Core Library barrel.
- */
 export * from './lib/mocks-core.module';
 export * from './lib/data-access/index';
 export * from './lib/ui/index';
