+++ libs/mocks-core/src/lib/data-access/http-handlers/http-handlers-provider.mock.ts
@@ -8,13 +8,13 @@ import {
   toasterServiceProvider,
   WebEnvironment,
 } from '@nx-ng-starter/shared-core/data-access';
+import { getWindow, WINDOW } from '@nx-ng-starter/shared-core/util';
 import {
   HttpProgressService,
   httpProgressServiceProvider,
   UserService,
   userServiceProvider,
-} from '@nx-ng-starter/shared-core/ui';
-import { getWindow, WINDOW } from '@nx-ng-starter/shared-core/util';
+} from '@nx-ng-starter/shared-store/state';
 import { HttpLink } from 'apollo-angular-link-http';
 
 import { dialogRefMockProvider } from '../../util/refs/dialog-ref.mock';
