+++ libs/mocks-core/src/lib/data-access/index.ts
@@ -1,4 +1,1 @@
-/**
- * Mocks Core Data Access barrel.
- */
 export * from './http-handlers/http-handlers-provider.mock';
