+++ libs/shared-core/src/index.ts
@@ -1,4 +1,1 @@
-/**
- * Shared Core Library barrel.
- */
 export * from './lib/shared-core.module';
