--- libs/shared-core/src/lib/ui/components/index.ts
@@ -1,4 +1,0 @@
-/**
- * Shared Core Library components barrel.
- */
-export const TODO = 'export components';
