+++ libs/shared-core/src/lib/ui/index.ts
@@ -1,6 +1,2 @@
-/**
- * Chared Core Library UI barrel.
- */
 export * from './animations/index';
-export * from './components/index';
 export * from './modules/index';
