+++ libs/shared-core/src/lib/ui/modules/index.ts
@@ -1,9 +1,2 @@
-/**
- * Shared Core Library modules barrel.
- */
 export * from './custom-material/custom-material.module';
 export * from './translate/app-translate.module';
-export * from './state/http-progress';
-export * from './state/http-api';
-export * from './state/sidebar-ui';
-export * from './state/user';
