+++ libs/shared-store/.eslintrc
@@ -1,0 +1,3 @@
+{
+  "extends": ["../../.eslintrc.js", "../../.eslintrc.angular.js"]
+}
