+++ libs/shared-store/README.md
@@ -1,0 +1,7 @@
+# shared-store
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-store` to execute the unit tests.
