+++ libs/shared-store/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/shared-store.module';
