+++ libs/shared-store/src/lib/state/http-api/http-api.interface.ts
