+++ libs/shared-store/src/lib/state/http-progress/http-progress.interface.ts
