+++ libs/shared-store/src/lib/state/http-progress/http-progress.module.ts
@@ -1,15 +1,14 @@
 import { ModuleWithProviders, NgModule, Provider } from '@angular/core';
 import { NgxsModule } from '@ngxs/store';
 
-import { httpProgressServiceProvider } from './http-progress.service';
+import { HttpProgressService } from './http-progress.service';
 import { HttpProgressState } from './http-progress.store';
 
-export const httpProgressModuleProviders: Provider[] = [httpProgressServiceProvider];
+export const httpProgressModuleProviders: Provider[] = [HttpProgressService];
 
 @NgModule({
   declarations: [],
   imports: [NgxsModule.forFeature([HttpProgressState])],
-  providers: [...httpProgressModuleProviders],
 })
 export class HttpProgressModule {
   public static forRoot(): ModuleWithProviders<HttpProgressModule> {
