+++ libs/shared-store/src/lib/state/sidebar-ui/sidebar-ui.interface.ts
