+++ libs/shared-store/tsconfig.lib.json
