+++ nx.json
@@ -29,6 +29,9 @@
     },
     "shared-core": {
       "tags": ["scope:shared-core", "type:feature", "platform:desktop"]
+    },
+    "shared-store": {
+      "tags": []
     }
   },
   "tasksRunnerOptions": {
