+++ package.json
@@ -108,15 +108,15 @@
   },
   "dependencies": {
     "@angular/animations": "10.0.1",
-    "@angular/cdk": "10.0.0",
+    "@angular/cdk": "10.0.1",
     "@angular/common": "10.0.1",
     "@angular/compiler": "10.0.1",
     "@angular/core": "10.0.1",
     "@angular/elements": "10.0.1",
     "@angular/flex-layout": "10.0.0-beta.32",
     "@angular/forms": "10.0.1",
-    "@angular/material": "10.0.0",
-    "@angular/material-moment-adapter": "10.0.0",
+    "@angular/material": "10.0.1",
+    "@angular/material-moment-adapter": "10.0.1",
     "@angular/platform-browser": "10.0.1",
     "@angular/platform-browser-dynamic": "10.0.1",
     "@angular/router": "10.0.1",
@@ -127,7 +127,7 @@
     "@nestjs/graphql": "7.4.6",
     "@nestjs/jwt": "7.0.0",
     "@nestjs/microservices": "7.2.0",
-    "@nestjs/passport": "7.0.0",
+    "@nestjs/passport": "7.1.0",
     "@nestjs/platform-express": "7.2.0",
     "@ngneat/until-destroy": "7.3.0",
     "apollo-angular": "1.9.1",
@@ -136,13 +136,13 @@
     "apollo-client": "2.6.10",
     "apollo-link": "1.2.14",
     "apollo-link-error": "1.1.13",
-    "apollo-server-express": "2.15.0",
+    "apollo-server-express": "2.15.1",
     "apollo-upload-client": "13.0.0",
     "apollo-utilities": "1.3.4",
     "class-transformer": "0.2.3",
     "class-validator": "0.12.2",
     "dotenv": "8.2.0",
-    "graphql": "15.1.0",
+    "graphql": "15.2.0",
     "graphql-subscriptions": "1.1.0",
     "graphql-tag": "2.10.3",
     "graphql-tools": "6.0.11",
@@ -176,10 +176,10 @@
     "@angular/language-service": "10.0.1",
     "@commitlint/cli": "9.0.1",
     "@commitlint/config-conventional": "9.0.1",
-    "@graphql-codegen/cli": "1.15.4",
-    "@graphql-codegen/typescript": "1.15.4",
-    "@graphql-codegen/typescript-apollo-angular": "1.15.4",
-    "@graphql-codegen/typescript-operations": "1.15.4",
+    "@graphql-codegen/cli": "1.16.0",
+    "@graphql-codegen/typescript": "1.16.0",
+    "@graphql-codegen/typescript-apollo-angular": "1.16.0",
+    "@graphql-codegen/typescript-operations": "1.16.0",
     "@nestjs/schematics": "7.0.1",
     "@nestjs/testing": "7.2.0",
     "@ngx-translate/core": "13.0.0",
@@ -187,15 +187,15 @@
     "@ngxs/form-plugin": "3.6.2",
     "@ngxs/logger-plugin": "3.6.2",
     "@ngxs/store": "3.6.2",
-    "@nrwl/angular": "9.4.5",
-    "@nrwl/cypress": "9.4.5",
-    "@nrwl/eslint-plugin-nx": "9.4.5",
-    "@nrwl/jest": "9.4.5",
-    "@nrwl/nest": "9.4.5",
-    "@nrwl/node": "9.4.5",
-    "@nrwl/storybook": "9.4.5",
-    "@nrwl/web": "9.4.5",
-    "@nrwl/workspace": "9.4.5",
+    "@nrwl/angular": "9.5.0",
+    "@nrwl/cypress": "9.5.0",
+    "@nrwl/eslint-plugin-nx": "9.5.0",
+    "@nrwl/jest": "9.5.0",
+    "@nrwl/nest": "9.5.0",
+    "@nrwl/node": "9.5.0",
+    "@nrwl/storybook": "9.5.0",
+    "@nrwl/web": "9.5.0",
+    "@nrwl/workspace": "9.5.0",
     "@storybook/addon-knobs": "5.3.19",
     "@storybook/angular": "5.3.19",
     "@types/apollo-upload-client": "8.1.3",
@@ -206,8 +206,8 @@
     "@types/jsdom": "16.2.3",
     "@types/marked": "1.1.0",
     "@types/node": "14.0.14",
-    "@typescript-eslint/eslint-plugin": "3.4.0",
-    "@typescript-eslint/parser": "3.4.0",
+    "@typescript-eslint/eslint-plugin": "3.5.0",
+    "@typescript-eslint/parser": "3.5.0",
     "@webcomponents/custom-elements": "1.4.1",
     "@webcomponents/webcomponentsjs": "2.4.3",
     "babel-loader": "8.1.0",
@@ -225,7 +225,7 @@
     "eslint-plugin-simple-import-sort": "5.0.3",
     "eslint-watch": "6.0.1",
     "firebase": "7.15.5",
-    "firebase-admin": "8.12.1",
+    "firebase-admin": "8.13.0",
     "firebase-functions": "3.7.0",
     "firebase-tools": "8.4.3",
     "google-protobuf": "3.12.2",
