+++ tools/shell/module-aliases.sh
@@ -19,6 +19,7 @@ MODULE_ALIAS_APP_NX_NG_STARTER_E2E="app:nx-ng-starter-e2e"
 MODULE_ALIAS_LIB_API_INTERFACE="lib:api-interface"
 MODULE_ALIAS_LIB_MOCKS_CORE="lib:mocks-core"
 MODULE_ALIAS_LIB_SHARED_CORE="lib:shared-core"
+MODULE_ALIAS_LIB_SHARED_STORE="lib:shared-store"
 MODULE_ALIAS_LIB_PROTO="lib:proto"
 
 MODULE_ALIAS_NAMES=(
@@ -28,6 +29,7 @@ MODULE_ALIAS_NAMES=(
   MODULE_ALIAS_LIB_API_INTERFACE
   MODULE_ALIAS_LIB_MOCKS_CORE
   MODULE_ALIAS_LIB_SHARED_CORE
+  MODULE_ALIAS_LIB_SHARED_STORE
   MODULE_ALIAS_LIB_PROTO
 )
 
@@ -38,6 +40,7 @@ MODULE_ALIAS_VARS=(
   "$MODULE_ALIAS_LIB_API_INTERFACE"
   "$MODULE_ALIAS_LIB_MOCKS_CORE"
   "$MODULE_ALIAS_LIB_SHARED_CORE"
+  "$MODULE_ALIAS_LIB_SHARED_STORE"
   "$MODULE_ALIAS_LIB_PROTO"
 )
 
@@ -47,6 +50,7 @@ MODULE_ALIAS_NAMES_UNIT=(
   MODULE_ALIAS_LIB_API_INTERFACE
   MODULE_ALIAS_LIB_MOCKS_CORE
   MODULE_ALIAS_LIB_SHARED_CORE
+  MODULE_ALIAS_LIB_SHARED_STORE
   MODULE_ALIAS_LIB_PROTO
 )
 
@@ -56,6 +60,7 @@ MODULE_ALIAS_VARS_UNIT=(
   "$MODULE_ALIAS_LIB_API_INTERFACE"
   "$MODULE_ALIAS_LIB_MOCKS_CORE"
   "$MODULE_ALIAS_LIB_SHARED_CORE"
+  "$MODULE_ALIAS_LIB_SHARED_STORE"
   "$MODULE_ALIAS_LIB_PROTO"
 )
